Add github actions for pull requests
This commit adds a github action that runs whenever a pull request
is opened, reopened, or a commit is added/updated. It runs go mod,
go fmt, a static checker test, go test, and checks that make manifests
produces no changes. It also ensures that the Dockerfile can be built.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>