Add Realese CRD and controller
This change adds the new Release CRD which will keep track of releases
statuses. The commit includes changes to the previous controller,
spreading the old functionality and the new one into two different
controllers.

The first one, pipelinerun_controller watches PipelineRuns both from the
Build and the Release service. It is in charge of creating new Release
objects and of updating their status.

The second one, release_controller, is quite similar to the one we had
before. Once a Release object is detected, it will create a new
PipelineRun to trigger the release.