Make controllers idempotent
With this change, the controllers are idempotent. To achieve it I have
added guards and, specially, I moved all the Release status updates to a
single controller by properly using Watches.

This change also adds a finalizer to the Release objects which is
invoked once those resources are marked to be deleted. The finalizer
removes all the resources "owned" by the Release. So far this is only
the PipelineRun created to trigger the Release Pipeline.