feat(HACBS-343): make operator idempotent
This change completely rewrites the operator, focusing on readability
and making it easier to be expanded in the future. The whole reconcile
loop for Releases has being reviewed. Now, watches (with proper
predicates) and owner objects are used to reduce the number of jumps to
different controllers and, specially, the number of reconciles.

The PipelineRun controller has being removed as well as part of this
change as it won't be part of M5.

Signed-off-by: David Moreno García <damoreno@redhat.com>