fix(HACBS-510): fail on description starting with uppercase
We are following Conventional Commits and using Gitlint to enforce the
specification. Although the description doesn't have a set-on-stone
format, examples on Conventional Commits spec seem to use always a
sentence starting with lowercase. This is not enforced by Gitlint, but
I think it would be good to enforce it. First because all our commits
will follow the same convention. Second because a new user following the
spec will already think this is the proper way.

Signed-off-by: David Moreno García <damoreno@redhat.com>