feat(HACBS-603): pass a workspace to the Release pipeline
The release pipeline is going to need a workspace to be able to share
information between tasks.

This commit adds two new variables that can be set to define the
default PersistentVolumeClaim and the default name for the workspace.

In addition, the ReleaseStrategy has now a new field to allow to
override the PVC.

Signed-off-by: David Moreno García <damoreno@redhat.com>