feat(HACBS-555): add auto-release label to ReleaseLinks
This commit adds a mutating webhook for ReleaseLink CRs. The webhook
checks to see if the release.appstudio.openshift.io/auto-release label
exists. If it does not exist, it adds it with a value of true. If it
does exist (meaning the user added the label when creating the CR),
it will do nothing.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>