feat(HACBS-965): use Patch instead of Update
Although it hasn't being a problem until now, with the addition of
tests we have started to see problems raised by the update of
resources. This could be solved by fixing the tests, but it's a good
moment to move to using Patch as it was suggested by Brian Cook.

Signed-off-by: David Moreno García <damoreno@redhat.com>