feat(HACBS-980): refactor Release status errors
This commit adds two new types of Release status errors. They are
called ReleasePlanValidationError and TargetDisabledError. They are
used let the user know that there is an issue with the ReleasePlans or
that the ReleasePlanAdmission is set to disabled, respectively. It
does not make sense to call this a ReleaseValidationError as is
currently the case, since there is nothing invalid about the Release
itself.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>