feat: add KCP sample yamls
This commit adds some resource yaml files to the config/samples/kcp
directory. These are for showing how to have the controller
reconcile a simple release in a KCP environment that includes both
a dev and managed workspace. They do not use permission claims,
but may be updated to at a later date.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>