fix: use proper context
Two of the context used in functions loading resources from the
cluster were using the wrong context. As a matter of fact, this
is only true given the way in which load the Application, but the
function loading the Application components should be more
specific and provide a way to indicate the context as the
Application passed as a parameter could come from the dev workspace.
This will be addressed once we introduce a `loader.go` file where
we will have functions to load resources from specific contexts.

Signed-off-by: David Moreno García <damoreno@redhat.com>