fix(HACBS-1700): return ReleasePlanAdmission with matching application
If there are two ReleasePlanAdmissions in the managed namespace with the
same origin, but different applications, the one with the matching
application should be returned, not just index 0 of the list of
ReleasePlanAdmissions found in the managed namespace that have a
matching origin.

This commit also has the adapter set the Release CR to a validation
error if there are two ReleasePlanAdmissions in a managed namespace with
the same origin and application.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>