feat(HACBS-2129): add release reference as a Pipeline parameter
This change adds a new function to declare new Pipeline parameters.
This function allows to pass any number of client.Objects. The kind
of the object will be used as the parameter name while the value will
contain the namespaced name of the object.

In addition, every new ReleasePipelineRun will contain a reference to
the Release.

Signed-off-by: David Moreno García <damoreno@redhat.com>