feat(HACBS-2126): create author webhook
This commit creates an author webhook. This author webhook operates on
Releases and ReleasePlans. It reverts the previous changes to the
Release webhook that accounted for author functionality; it is all in
the author webhook now. The author webhook is responsible for adding an
author label to Releases and ReleasePlans when they are created or
updated.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>