fix(RHTAPBUGS-560): ensure each pipelinerun uses its own subdir
When release pipelineruns are created, they will always use
the same shared workspace (provided by the default PVC, or one
defined in the given ReleaseStrategy). This wasn't an issue before,
but now that we started using the workspace to share snapshot and
other data from task to task, we found out that two simultaneously
running pipelineruns can alter each other's data. We need to prevent
that.

This typically happens when you have two or more components in your
app and so there are multiple pipeline runs, oftentimes running
around the same time.

The solution is to use a subdir when passing the PVC in the PR.
The subdir is based on the Release name plus the PR's UID,
so it should always be unique.

Signed-off-by: Martin Malina <mmalina@redhat.com>