fix(RHTAPBUGS-691): revert subPath creation in PRs
Recently we made a change to start using subPath
when defining a workspace during PipelineRun creation.

This was to fix an issue when 2 or more PRs running
in parallel would override each other's data.

It turns out that there is a race condition happening often
when kubernetes tries to create the subPath in the volume.
When there are several pods with the same subPath created
at the same time (as is the case with Tekton PRs - there are
pods for each task), it can happen that several pods try
to create the subPath and only one of them succeeds. See jira
for more details.

We'll have to find another solution to the problem we were
fixing with this. In the meantime, we'll have to revert this.

This reverts commit 25cfcdb12688b0b679cd4beb923662e580ef2da3.

Signed-off-by: Martin Malina <mmalina@redhat.com>