fix: update filters and panels
The Release Grafana dashboard was not working properly. I found out
that some filters were not properly defined. I also saw that one of
the panels was deprecated, so I updated it.

In any case, there seems to be an unrelated problem. Just editing
the panels and clicking on the query was enough for the panel to
start showing values (without any changes). Let's see how it behaves
after this change.

Signed-off-by: David Moreno García <damoreno@redhat.com>