fix: modify gosec PR check to log results
A while ago we modified the workflow to save the results
in a SARIF file and then upload to Github to show issues
in the UI, but somehow it does not seem to work right now
and there is no way to see what's wrong.

This change adds a second run of gosec which will
print the issues in the log directly.

Signed-off-by: Martin Malina <mmalina@redhat.com>