fix(RHTAPREL-854): add processing cleanup operation to Release adapter
The Release Processing resources were cleaned up as part of the
operation that tracks the Processing status and if an error occurred
cleaning them up, it was not retried. This commit breaks the cleanup out
into its own operation that will run once the Release is marked as
Processed and properly retries if an error occurs during cleanup.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>