fix: remove pattern
The release-service deployment is CrashLooping right now. It is
possible that this is caused by multiple RP/RPAs being reconciled
because an error that arises whenever the name of a matched RPA is
set to empty string. This is a quick fix to remove the pattern check
and see if this improves the situation.

Signed-off-by: David Moreno García <damoreno@redhat.com>