feat(RELEASE-1227): set application metadata in release controller
This commit adds an operation to the release controller to set metadata
from the application on the release. It sets the ownerReference to be
the application as well as copies the pipelines as code and rhtapDomain
labels and annotations from the snapshot to the release. This is
currently done for automated releases by the integration service, so
this commit moves the logic to the release controller instead.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>