feat(RELEASE-1434): allow to override volume type (#666)
This is a provisional change that will be used during the
transition to using Trusted Artificts in the
release-service-catalog pipelines.

Trusted Artifacts requireis to use an emptyDir in opposition
to the PVCs used today. This change allow to override
the volume type used by a given pipeline by defining a list
of EmptyDirOverride items which contain the details of the
pipeline to be overriden. Regex is supported in the url and
revision values.

Signed-off-by: David Moreno García <damoreno@redhat.com>