fix(KONFLUX-7965): reconcile loop stuck when tenant pipeline fails
When a tenant pipeline fails, the managed pipeline never gets
created, and consequently the cleanup logic that waits for
all pipelines to finish gets stuck in an infinite loop.

This fix ensures that the managed pipeline gets skipped
if the tenant pipeline fails.

Also added new methods for checking if a pipeline is skipped.

Signed-off-by: Filip Nikolovski <fnikolov@redhat.com>