refactor(RELEASE-1352): add block-releases label to RPAs
This commit is the first for RELEASE-1352. It adds the block-releases
label for ReleasePlanAdmissions. This will take the place of the
auto-release label, with the hope of increasing clarity. The
auto-release label still exists with this commit, but once all
ReleasePlanAdmissions are migrated, a future commit will remove it (just
for RPAs, not for RPs).

Signed-off-by: Johnny Bieren <jbieren@redhat.com>