feat(KONFLUX-8553): support cel expressions in auto-release label
Users have requested more granular control over whether a snapshot is
auto-released.  The integration service has added logic to support cel
expressions in the release.appstudio.openshift.io/auto-release label.
This PR removes the logic from the RelasePlan webhook that requires the
value of the label to be 'true' or 'false'

Signed-off-by: Ryan Cole <rcyoalne@gmail.com>