fix(RELEASE-1902): fix gosec CI for merge queues
The upload of SARIF file fails when the CI runs for merge queues. The
error is "Resource not accessible by integration", and is likely caused
by insufficient permissions when triggered by a merge queue. Resolve it
by skipping this step when triggered by a merge queue. The upload
doesn't have much value in this case, since it was already uploaded on
the pull request event.

Signed-off-by: Lubomir Gallovic <lgallovi@redhat.com>