fix(RELEASE-1902): fix e2e tests on merge queues
The reason e2e tests fail on merge queue is due to missing image. The
image is missing because the event is a "push" and thus test-metadata
step expects the image tag to not have the "on-pr-" prefix. Images
without this prefix are only created with the push pipeline, which means
that this pipeline also has to be triggered on merge queue events.

Signed-off-by: Lubomir Gallovic <lgallovi@redhat.com>